refactor Readers
