+++ ReadingOutputFromFLYonPIC_v2/reader/__init__.py
@@ -9,8 +9,17 @@ Authors: Brian Edward Marre
 License: GPLv3+
 """
 
-from . import openPMD
-from . import SCFLY
-from . import TimingData
+from . import Reader
+from . import StateDistributionReader
+from . import OpenPMDParticleReader
+from . import OpenPMDBinningReader
+from . import SCFLYReader
+from . import TimingDataReader
 
-__all__ = ["openPMD", "SCFLY", "TimingData"]
+__all__ = [
+    "Reader",
+    "StateDistributionReader",
+    "OpenPMDParticleReader",
+    "OpenPMDBinningReader",
+    "SCFLYReader",
+    "TimingDataReader"]
